fix: transaction logic

+++ src/controllers/transactionController.ts
@@ -1,17 +1,16 @@
-import { Request, Response } from 'express';
-import { Wallet, Transaction } from '../models';
+import e, { Request, Response } from 'express';
+import { Wallet, Transaction } from '@/models';
+import { WalletService } from '@/services';
+import { convertToDate } from '@/utils';
 
 export const getAllTransactions = async (req: Request, res: Response) => {
   try {
     const { query } = req;
-    console.log(query);
 
     const page = (parseInt(query.page as string) || 1) - 1;
     const perPage = parseInt(query.perPage as string) || 10;
 
-    const transactions = await Transaction.query()
-      // .where('tag', '=', 'salary')
-      .page(page, perPage);
+    const transactions = await Transaction.query().page(page, perPage);
 
     res.json(transactions);
   } catch (error) {
@@ -22,9 +21,16 @@ export const getAllTransactions = async (req: Request, res: Response) => {
 
 export const getTransactionsByTag = async (req: Request, res: Response) => {
   try {
-    const tag = req.params.tag;
+    const { query } = req;
+    const { tag } = req.params;
 
-    const transactions = await Transaction.query().where('tag', tag);
+    const page = (parseInt(query.page as string) || 1) - 1;
+    const perPage = parseInt(query.perPage as string) || 10;
+
+    const transactions = await Transaction.query()
+      .page(page, perPage)
+      .where('tag', tag);
+
     res.json(transactions);
   } catch (error) {
     console.error(error);
@@ -32,22 +38,66 @@ export const getTransactionsByTag = async (req: Request, res: Response) => {
   }
 };
 
+export const deleteTransactionTag = async (req: Request, res: Response) => {
+  try {
+    const { id } = req.params;
+
+    // Check if the transaction exists
+    const transaction = await Transaction.query().findById(id);
+
+    if (!transaction) {
+      res.status(404).json({ message: 'Transaction not found' });
+      return;
+    }
+
+    // Update the transaction in the database
+    const updatedTransaction = await Transaction.query().updateAndFetchById(
+      id,
+      { ...transaction, tag: '' }
+    );
+
+    if (updatedTransaction) {
+      res.json(updatedTransaction);
+    } else {
+      res.status(404).json({ message: 'Transaction not updated' });
+    }
+  } catch (error: any) {
+    console.error(error);
+    res.status(500).json({
+      message: error.nativeError?.sqlMessage || error.message
+    });
+  }
+};
 export const getTransactionsByWallet = async (req: Request, res: Response) => {
   try {
-    const walletId = req.params.walletId;
-    const transactions = await Transaction.query().where('wallet_id', walletId);
+    const { query } = req;
+    const { wallet_id } = req.params;
+
+    const page = (parseInt(query.page as string) || 1) - 1;
+    const perPage = parseInt(query.perPage as string) || 10;
+
+    const transactions = await Transaction.query()
+      .page(page, perPage)
+      .where('wallet_id', wallet_id);
+
     res.json(transactions);
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: 'Error fetching transactions by wallet' });
   }
 };
-
 export const getTransactionsByType = async (req: Request, res: Response) => {
   try {
     const { type } = req.params;
+    const { query } = req;
 
-    const transactions = await Transaction.query().where('type', type);
+    const page = (parseInt(query.page as string) || 1) - 1;
+    const perPage = parseInt(query.perPage as string) || 10;
+
+    const transactions = await Transaction.query()
+      .page(page, perPage)
+      .where('type', type);
+
     res.json(transactions);
   } catch (error) {
     console.error(error);
@@ -71,24 +121,25 @@ export const getTransactionById = async (req: Request, res: Response) => {
   }
 };
 
-function convertToDate(dateStrs: string[]) {
-  return dateStrs.map(dateStr => new Date(dateStr));
-}
-
 export const getStatistics = async (req: Request, res: Response) => {
   try {
-    const { range: datesRange } = req.body;
+    const { startRange, endRange } = req.body;
+    const { query } = req;
 
-    const dates = convertToDate(datesRange);
+    const page = (parseInt(query.page as string) || 1) - 1;
+    const perPage = parseInt(query.perPage as string) || 10;
 
-    const transactions = await Transaction.query().whereBetween('date_added', [dates[0], dates[1]]);
+    const dates = convertToDate([startRange, endRange]);
+
+    const transactions = await Transaction.query()
+      .page(page, perPage)
+      .whereBetween('date_added', [dates[0], dates[1]]);
 
     if (!transactions) {
       res.status(404).json({ message: 'Transaction not found' });
     } else {
       res.json(transactions);
     }
-
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: 'Error fetching transaction' });
@@ -107,47 +158,90 @@ export const createTransaction = async (req: Request, res: Response) => {
       return;
     }
 
-    // Calculate the new wallet balance
-    const resultByType: Record<string, number> = {
-      income: wallet.balance + +amount,
-      expense: wallet.balance - +amount,
-    };
-    const newBalance = resultByType[type];
+    // Add the new transaction to the database
+    const transaction = await Transaction.query().insert({
+      ...req.body,
+      type,
+      wallet_id: +wallet.id,
+      amount: +amount
+    });
 
     // Update the wallet balance
-    await Wallet.query().updateAndFetchById(wallet_id, { balance: newBalance });
+    const walletService = new WalletService();
+    await walletService.updateBalance(wallet, transaction, 'add');
 
-    // Add the new transaction to the database
-    const transaction = await Transaction.query().insert({ ...req.body, wallet_id: +wallet.id, amount: +amount });
     res.json(transaction);
-  } catch (error) {
+  } catch (error: any) {
     console.error(error);
-    res.status(500).json({ message: 'Error creating transaction' });
+    res.status(500).json({
+      message: error.nativeError?.sqlMessage || error.message
+    });
   }
 };
 
 export const updateTransaction = async (req: Request, res: Response) => {
   try {
     const id = req.params.id;
-    const transaction = await Transaction.query().update({ ...req.body }).where('id', id);
+
+    // Check if the transaction exists
+    const transaction = await Transaction.query().findById(id);
+
     if (!transaction) {
       res.status(404).json({ message: 'Transaction not found' });
-    } else {
-      res.json(transaction);
+      return;
     }
-  } catch (error) {
+
+    // Update the transaction in the database
+    const updatedTransaction = await Transaction.query().updateAndFetchById(
+      id,
+      { ...transaction, ...req.body }
+    );
+
+    if (updatedTransaction) {
+      res.json(updatedTransaction);
+    } else {
+      res.status(404).json({ message: 'Transaction not updated' });
+    }
+  } catch (error: any) {
     console.error(error);
-    res.status(500).json({ message: 'Error updating transaction' });
+    res.status(500).json({
+      message: error.nativeError?.sqlMessage || error.message
+    });
   }
 };
 
 export const deleteTransaction = async (req: Request, res: Response) => {
   try {
-    const id = req.params.id;
+    const { id } = req.params;
+
+    // Check if the transaction exists
+    const transaction = await Transaction.query().findById(id);
+
+    if (!transaction) {
+      res.status(404).json({ message: 'Transaction not found' });
+      return;
+    }
+
+    // Check if the wallet exists
+    const wallet = await Wallet.query().findById(transaction.wallet_id);
+
+    if (!wallet) {
+      res.status(404).json({ message: 'Wallet not found' });
+      return;
+    }
+
+    // Update the wallet balance
+    const walletService = new WalletService();
+    await walletService.updateBalance(wallet, transaction, 'delete');
+
+    // Delete the transaction from the database
     await Transaction.query().deleteById(id);
+
     res.json({ message: 'Transaction deleted' });
-  } catch (error) {
+  } catch (error: any) {
     console.error(error);
-    res.status(500).json({ message: 'Error deleting transaction' });
+    res
+      .status(500)
+      .json({ message: error.nativeError?.sqlMessage || error.message });
   }
 };
